+++ src/lib/documents/Get-CosmosDbDocument.ps1
@@ -43,7 +43,7 @@ function Get-CosmosDbDocument
 
         [Parameter()]
         [ValidateNotNullOrEmpty()]
-        [System.String[]]
+        [System.Object[]]
         $PartitionKey,
 
         [Parameter()]
@@ -178,7 +178,7 @@ function Get-CosmosDbDocument
         if ($PSBoundParameters.ContainsKey('PartitionKey'))
         {
             $headers += @{
-                'x-ms-documentdb-partitionkey' = '["' + ($PartitionKey -join '","') + '"]'
+                'x-ms-documentdb-partitionkey' = Format-CosmosDbDocumentPartitionKey -PartitionKey $PartitionKey
             }
             $null = $PSBoundParameters.Remove('PartitionKey')
         }